Add delete_folder function to S3
There are a lot of missing or failing unit tests. I've been doing a
sort of "test-assisted" development. So I will use cucumber end-to-end
tests to drive the programming as the whole point is for this to
interface with cloud services and I have to learn how they work as I go.
I'll then go back and add unit tests while cleaning up the code.