Integrate S3 (#6)
* Start with S3 and add upload and delete file

* Fix S3 integration unit test

* Add S3 download_file function

* Fix S3 cucumber credentials setup

* Add S3 list_files

* Add S3 list_all_files but no unit tests yet

* Add unit tests for list_all_files

* Add get_file_info to S3

* Start S3 File and Folder Abstractions.

I've realized now that I will actually need to make a delete_folder
method for S3 that recurses through and deletes sub-folders and
sub-directories.

* Add delete_folder function to S3

There are a lot of missing or failing unit tests. I've been doing a
sort of "test-assisted" development. So I will use cucumber end-to-end
tests to drive the programming as the whole point is for this to
interface with cloud services and I have to learn how they work as I go.
I'll then go back and add unit tests while cleaning up the code.

* Fix S3 list_all_files tests

* Add unit tests for S3 delete_folder

* Add tests for S3 File and Folder abstractions

* Fix some not-synchronous issues with download_file

* Add S3 to done on README table